requirements.txt
Flask==1.1.2
matplotlib==3.4.3
numpy==1.20.3
opencv_python==4.5.5.64
Pillow==9.1.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.62.3
Werkzeug==2.0.2
